Code cleanup for jose-util
