+++ jose-util/jose-util.t
@@ -92,3 +92,17 @@ Expand a compact message to full format.
   $ echo "eyJhbGciOiJFUzM4NCJ9.TG9yZW0gaXBzdW0gZG9sb3Igc2l0IGFtZXQK.QPU35XY913Im7ZEaN2yHykfbtPqjHZvYp-lV8OcTAJZs67bJFSdTSkQhQWE9ch6tvYrj_7py6HKaWVFLll_s_Rm6bmwq3JszsHrIvFFm1NydruYHhvAnx7rjYiqwOu0W" |
   > jose-util expand --format JWS
   {"payload":"TG9yZW0gaXBzdW0gZG9sb3Igc2l0IGFtZXQK","protected":"eyJhbGciOiJFUzM4NCJ9","signature":"QPU35XY913Im7ZEaN2yHykfbtPqjHZvYp-lV8OcTAJZs67bJFSdTSkQhQWE9ch6tvYrj_7py6HKaWVFLll_s_Rm6bmwq3JszsHrIvFFm1NydruYHhvAnx7rjYiqwOu0W"}
+
+Base64-decode data in various formats (padded, unpadded, standard, url-safe).
+
+  $ echo "8J+Ukgo=" | jose-util b64decode
+  🔒
+
+  $ echo "8J+Ukgo" | jose-util b64decode
+  🔒
+
+  $ echo "8J-Ukgo=" | jose-util b64decode
+  🔒
+
+  $ echo "8J-Ukgo" | jose-util b64decode
+  🔒
