--- jose-util/test-keys/ecdh.key
@@ -1,6 +1,0 @@
------BEGIN EC PRIVATE KEY-----
-MIGkAgEBBDDvoj/bM1HokUjYWO/IDFs26Jo0GIFtU3tMQQu7ZabKscDMK3dZA0mK
-v97ij7BBFbCgBwYFK4EEACKhZANiAAT3KhQQCDFN32y/B72g+qOFw/5/aNx1MvZa
-rwDDa/2G3V0HLTS0VE82sLEUKS8xwkWFI+gNRXk0vvN+Hf+myJI1jOIY+tYQlh+C
-ZiKGNJ6g5/Su7V6ukGtN+UiY+sx+0LI=
------END EC PRIVATE KEY-----
