--- jose-util/test-keys/ecdh.pub
@@ -1,5 +1,0 @@
------BEGIN PUBLIC KEY-----
-MHYwEAYHKoZIzj0CAQYFK4EEACIDYgAE9yoUEAgxTd9svwe9oPqjhcP+f2jcdTL2
-Wq8Aw2v9ht1dBy00tFRPNrCxFCkvMcJFhSPoDUV5NL7zfh3/psiSNYziGPrWEJYf
-gmYihjSeoOf0ru1erpBrTflImPrMftCy
------END PUBLIC KEY-----
